refactor: phone_adapter 文件拆分 + ncp_dispatch 路由分离 + 代码内聚
- 新建 ncp_dispatch.cpp: 路由层 (g_isPhone + wrapper), 独立于 phone_adapter
- phone_adapter/ 按功能拆分为:
  phone_socket.h (header-only inline) — 公共 socket I/O, 消除重复
  phone_process.h/.cpp — fork 进程创建 (从 phone_adapter.cpp 移入)
  phone_virgl_relay.h/.cpp — virgl 中继 (从 graphics_broker.cpp 移入)
  phone_virgl_dispatch.h/.cpp — virgl dispatch (从 virgl_child.cpp 移入)
- graphics_broker.cpp: 删除 ~140 行 relay 代码, 改为调用 PhoneVirgl_RelayRequest()
- virgl_child.cpp: 删除 ~115 行 dispatch 代码, 改为调用 PhoneVirgl_DispatchStart()
- phone_adapter/ 现在是纯手机代码, 无路由分支

构建通过 (arm64-v8a)
