+++ entry/src/main/cpp/phone_adapter/phone_process.h
@@ -1,0 +1,16 @@
+// phone_process.h — 手机适配层：fork 进程创建
+#pragma once
+#include <AbilityKit/native_child_process.h>
+
+// fork 版进程创建（纯手机代码，无路由逻辑）
+Ability_NativeChildProcess_ErrCode Phone_StartNativeChildProcess(
+    const char* entry, NativeChildProcess_Args args,
+    NativeChildProcess_Options options, int32_t* pid);
+
+int Phone_CreateNativeChildProcess(
+    const char* libName, OH_Ability_OnNativeChildProcessStarted onProcessStarted);
+
+// Proxy 查询接口（graphics_broker.cpp 使用）
+bool PhoneAdapter_IsDummyProxy(const void* p);
+int  PhoneAdapter_GetConfigSocket(void);
+void PhoneAdapter_CloseConfigSocket(void);
