+++ entry/src/main/cpp/virgl_child.cpp
@@ -1,4 +1,5 @@
 #include <AbilityKit/native_child_process.h>
+#include "phone_adapter/phone_virgl_dispatch.h"
 #include <EGL/egl.h>
 #include <EGL/eglext.h>
 #include <GLES3/gl3.h>
@@ -262,121 +263,6 @@ int OnVtestPresent(uint32_t texId, uint32_t width, uint32_t height,
 
 extern "C" __attribute__((visibility("default"))) void Main(NativeChildProcess_Args args);
 
-// ==================== 手机适配层：virgl IPC dispatch ====================
-// fork 模式下没有 Binder 驱动回调 OnVirglIpcRequest；本线程从配置 socket
-// 读请求、按 virgl_ipc 协议构造 parcel、直接调用 OnVirglIpcRequest，并回送 reply。
-// 线协议与 graphics_broker.cpp 的 PhoneVirglRelayRequestLocked 严格对应。
-namespace {
-constexpr uint32_t kPhoneMaxPayload = 4096;
-bool PhoneSockReadAll(int fd, void* buf, size_t len) {
-    uint8_t* p = static_cast<uint8_t*>(buf);
-    while (len > 0) {
-        ssize_t n = recv(fd, p, len, 0);
-        if (n < 0) { if (errno == EINTR) continue; return false; }
-        if (n == 0) return false;
-        p += n; len -= static_cast<size_t>(n);
-    }
-    return true;
-}
-bool PhoneSockWriteAll(int fd, const void* buf, size_t len) {
-    const uint8_t* p = static_cast<const uint8_t*>(buf);
-    while (len > 0) {
-        ssize_t n = send(fd, p, len, MSG_NOSIGNAL);
-        if (n < 0) { if (errno == EINTR) continue; return false; }
-        p += n; len -= static_cast<size_t>(n);
-    }
-    return true;
-}
-template <typename T> bool PhoneSockReadPod(int fd, T& v) { return PhoneSockReadAll(fd, &v, sizeof(v)); }
-template <typename T> bool PhoneSockWritePod(int fd, const T& v) { return PhoneSockWriteAll(fd, &v, sizeof(v)); }
-bool PhoneSockReadStr(int fd, std::string& out) {
-    uint32_t len = 0;
-    if (!PhoneSockReadPod(fd, len) || len > 1024) return false;
-    if (len == 0) { out.clear(); return true; }
-    std::vector<char> buf(len);
-    if (!PhoneSockReadAll(fd, buf.data(), len)) return false;
-    out.assign(buf.data(), len - 1);   // 去掉 NUL 结尾
-    return true;
-}
-void PhoneVirglDispatchLoop(int fd) {
-    namespace vi = winehua::virgl_ipc;
-    OH_LOG_INFO(LOG_APP, "[PhoneVirgl] dispatch loop start fd=%{public}d pid=%{public}d",
-                fd, getpid());
-    while (true) {
-        uint32_t code = 0, len = 0;
-        if (!PhoneSockReadPod(fd, code) || !PhoneSockReadPod(fd, len) || len > kPhoneMaxPayload) break;
-        OHIPCParcel* data = OH_IPCParcel_Create();
-        OHIPCParcel* reply = OH_IPCParcel_Create();
-        bool ok = (data && reply);
-        if (ok) {
-            switch (code) {
-            case vi::kConfigureRequest: {
-                int32_t version = 0; std::string s[5];
-                ok = PhoneSockReadPod(fd, version);
-                for (int i = 0; ok && i < 5; ++i) ok = PhoneSockReadStr(fd, s[i]);
-                if (ok) {
-                    OH_IPCParcel_WriteInt32(data, version);
-                    for (int i = 0; i < 5; ++i) OH_IPCParcel_WriteString(data, s[i].c_str());
-                }
-                break;
-            }
-            case vi::kDetachSurfaceRequest: {
-                int32_t version = 0; int64_t key = 0;
-                ok = PhoneSockReadPod(fd, version) && PhoneSockReadPod(fd, key);
-                if (ok) {
-                    OH_IPCParcel_WriteInt32(data, version);
-                    OH_IPCParcel_WriteInt64(data, key);
-                }
-                break;
-            }
-            case vi::kSetFramePeriodRequest: {
-                int32_t version = 0; int64_t key = 0, period = 0;
-                ok = PhoneSockReadPod(fd, version) && PhoneSockReadPod(fd, key) && PhoneSockReadPod(fd, period);
-                if (ok) {
-                    OH_IPCParcel_WriteInt32(data, version);
-                    OH_IPCParcel_WriteInt64(data, key);
-                    OH_IPCParcel_WriteInt64(data, period);
-                }
-                break;
-            }
-            case vi::kShutdownRequest:
-            case vi::kQuerySurfacesRequest: {
-                int32_t version = 0;
-                ok = PhoneSockReadPod(fd, version);
-                if (ok) OH_IPCParcel_WriteInt32(data, version);
-                break;
-            }
-            default:
-                ok = false;
-            }
-        }
-        if (!ok) {
-            if (data) OH_IPCParcel_Destroy(data);
-            if (reply) OH_IPCParcel_Destroy(reply);
-            break;
-        }
-        // 与 Binder 驱动调用等价：直接调业务入口
-        OnVirglIpcRequest(code, data, reply, nullptr);
-        // 回送 reply：query=buffer；其余=int32 result
-        if (code == vi::kQuerySurfacesRequest) {
-            const uint8_t* bytes = OH_IPCParcel_ReadBuffer(
-                reply, static_cast<int32_t>(sizeof(vi::SurfaceQueryReply)));
-            uint32_t rlen = bytes ? static_cast<uint32_t>(sizeof(vi::SurfaceQueryReply)) : 0;
-            PhoneSockWritePod(fd, rlen);
-            if (rlen) PhoneSockWriteAll(fd, bytes, rlen);
-        } else {
-            int32_t result = -1;
-            OH_IPCParcel_ReadInt32(reply, &result);
-            uint32_t rlen = sizeof(result);
-            PhoneSockWritePod(fd, rlen);
-            PhoneSockWritePod(fd, result);
-        }
-        OH_IPCParcel_Destroy(data);
-        OH_IPCParcel_Destroy(reply);
-    }
-    OH_LOG_INFO(LOG_APP, "[PhoneVirgl] dispatch loop exit fd=%{public}d", fd);
-}
-} // namespace
 extern "C" __attribute__((visibility("default"))) OHIPCRemoteStub* NativeChildProcess_OnConnect()
 {
     g_virglIpcStub = OH_IPCRemoteStub_Create(
@@ -392,12 +278,11 @@ extern "C" __attribute__((visibility("default"))) void NativeChildProcess_MainPr
     setenv("EGL_PLATFORM", "surfaceless", 1);
     // 手机适配层：启动 socket dispatch 线程，替代 Binder 驱动回调
     {
-        const char* shimFdEnv = getenv("WINEHUA_PHONE_CFG_FD");
-        if (shimFdEnv && shimFdEnv[0]) {
-            int shimFd = atoi(shimFdEnv);
-            OH_LOG_INFO(LOG_APP, "[PhoneVirgl] phone mode, starting dispatch on fd=%{public}d",
-                        shimFd);
-            std::thread(PhoneVirglDispatchLoop, shimFd).detach();
+        const char* fdEnv = getenv("WINEHUA_PHONE_CFG_FD");
+        if (fdEnv && fdEnv[0]) {
+            int fd = atoi(fdEnv);
+            OH_LOG_INFO(LOG_APP, "[PhoneVirgl] phone mode, starting dispatch on fd=%{public}d", fd);
+            PhoneVirgl_DispatchStart(fd, OnVirglIpcRequest);  // → phone_adapter/phone_virgl_dispatch.cpp
         }
     }
     std::unique_lock<std::mutex> lock(g_ipcChildMutex);
